# index.js
const express = require('express');
require('dotenv').config();
const Question = require('./models/Question');
const Answer = require('./models/Answer');
const User = require('./models/User');
const Category = require('./models/category')
const auth = require('./middleware/auth');
const mongoose = require('mongoose');
const path = require('path');
const cors = require('cors');
const app = express();
const swaggerUi = require('swagger-ui-express');

const swaggerDocument = require('./swagger.json');

mongoose.connect(process.env.MONGODB_URL, {
    useNewUrlParser: true
})

app.use(express.json());

app.use(cors());

app.use(express.static(path.resolve(__dirname, "./react-app/build")));

app.use('/api-docs', swaggerUi.serve, swaggerUi.setup(swaggerDocument));

app.get('/', (req, res) => {
    res.send(path.join(__dirname, './react-app/build', 'index.js'))
})

app.post('/user/login', async (req, res) => {
    console.log(req.body)
    try {
        const user = await User.findByCredentials(req.body.email, req.body.password);
        const token = await user.generateAuthToken();
        res.status(200).send({ user, token });
    } catch (e) {
        console.log(e)
        res.status(400).send(e);
    }
});

app.post('/user/register', async (req, res) => {
    const user = new User(req.body);
    try {
        await user.save();
        const token = await user.generateAuthToken();
        res.status(201).send({ user, token });
    } catch (e) {
        res.status(400).send(e);
    }
})

app.post('/addcategory', auth, async (req, res) => {
    const category = new Category(req.body);
    try {
        await category.save();
        res.status(201).send();
    } catch (e) {
        res.status(400).send(e);
    }
})



app.get('/getCategory/:id', async (req, res) => {
    try {
        const category = await Category.findById(req.params.id);
        res.status(200).send({
            category: category.category
        })
    } catch (e) {
        res.status(400).send(e);
    }
})


app.get('/qa/getcategories', async (req, res) => {
    try {
        const categories = await Category.find();
        res.send({ categories });
    }
    catch (e) {
        res.status(400).send(e);
    }
})

app.get('/qa/getquestions', auth, async (req, res) => {
    try {
        const questions = await Question.find();
        res.status(200).send({
            questions
        })
    } catch (error) {
        res.status(400).send(error);
    }
    
})

app.post('/qa/postquestion', auth, async (req, res) => {
    const question = new Question(req.body);
    try {
        await question.save();
        return res.status(201).send();
    } catch (e) {
        console.log(e)
        res.status(404).send();
    }
});


app.post('/qa/postanswers', auth, async (req, res) => {
    console.log(req.body)
    const answer = new Answer(req.body);
    try {
        await answer.save();
        return res.status(201).send();
    } catch (e) {
        console.log(e)
        res.status(404).send();
    }
})
app.get('/qa/showAnswers', async (req, res) => {
    const user = await User.findOne({})
    const answers = await Answer.find({ user_id: user._id });
    const ans = []
    for (let i = 0; i < answers.length; i++) {
        const ques = await Question.findOne({ _id: answers[i].question_id })
        ans.push({ name: user.name, question: ques?.question, answer: answers[i].answer, createdAt: answers[i].createdAt })
        //console.log(Date(answers[i].createdAt))
    }
    res.send({ ans })
})



app.listen(process.env.PORT, () => {
    console.log("------------Server running on port ", process.env.PORT);
})